+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.0.5 (2017-04-09)
+- added whole word and ignore case
+- only update current document on content change
+- apply highlights on peek when initially inactive
+
 ## 0.0.4 (2017-01-14)
 - remove annoyance